--- frontend/src/components/Profile/ProfileCont/index.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-import PProfileCont from "./PProfileCont";
-
-export default function ProfileCont() {
-  return (
-    <>
-      <PProfileCont />
-    </>
-  );
-}
