+++ frontend/src/pages/Profile/index.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 
 import { NavBar } from "../../components/Common";
-import ProfileCont from "../../components/Profile/ProfileCont";
+import ProfileComponent from "../../components/Profile";
 
 export default function Profile() {
   return (
     <>
       <NavBar />
-      <ProfileCont />
+      <ProfileComponent />
     </>
   );
 }
